+++ src/game/jobs/Training.cpp
@@ -21,15 +21,252 @@
 #include "character/sGirl.h"
 #include "cGirls.h"
 #include "buildings/IBuilding.h"
+#include "utils/streaming_random_selection.hpp"
 
-struct sTrainingData {
-    const char* TargetTrait;
-    const char* HasAlreadyMessage;
-    const char* ProceedMessage;
-    const char* ResistMessage;
-    const char* ProgressMessage;
-    Image_Types ImageType;
-};
+extern const char* const TrainingInteractionId;
+
+namespace {
+    struct sTrainingData {
+        const char* TargetTrait;
+        const char* HasAlreadyMessage;
+        const char* ProceedMessage;
+        const char* ResistMessage;
+        const char* ProgressMessage;
+        Image_Types ImageType;
+    };
+
+    class PracticeJob : public cBasicJob {
+    public:
+        PracticeJob();
+
+        sWorkJobResult DoWork(sGirl& girl, bool is_night) override;
+        eCheckWorkResult CheckWork(sGirl& girl, bool is_night) override;
+    };
+
+    class MistressJob : public cBasicJob {
+    public:
+        MistressJob();
+
+        sWorkJobResult DoWork(sGirl& girl, bool is_night) override;
+        eCheckWorkResult CheckWork(sGirl& girl, bool is_night) override;
+    };
+}
+
+MistressJob::MistressJob() : cBasicJob(JOB_MISTRESS, "Mistress.xml") {
+    m_Info.FreeOnly = true;
+}
+
+sWorkJobResult MistressJob::DoWork(sGirl& girl, bool is_night) {
+    if (m_Performance > 150) {
+        girl.m_Building->ProvideInteraction(TrainingInteractionId, &girl, 2);
+        add_text("work.good");
+    } else {
+        add_text("work.normal");
+        girl.m_Building->ProvideInteraction(TrainingInteractionId, &girl, 1);
+    }
+
+    apply_gains(girl, m_Performance);
+
+    if(m_Performance > 100) {
+        auto* target_girl = girl.m_Building->girls().get_random_girl([](const sGirl& g) {
+            return g.is_slave() && g.obedience() < 50;
+        });
+        if (target_girl) {
+            SetSubstitution("slacker", target_girl->FullName());
+            add_text("slacker");
+            target_girl->upd_temp_stat(STAT_OBEDIENCE, 4);
+            target_girl->upd_temp_stat(STAT_PCFEAR, 2);
+        }
+    }
+
+    girl.AddMessage(ss.str(), IMGTYPE_TEACHER, is_night ? EVENT_NIGHTSHIFT : EVENT_DAYSHIFT);
+    return {false, 0, 0, 100};
+}
+
+IGenericJob::eCheckWorkResult MistressJob::CheckWork(sGirl& girl, bool is_night) {
+    if (girl.disobey_check(ACTION_WORKMATRON, JOB_MISTRESS))            // they refuse to work
+    {
+        add_text("refuse");
+        girl.AddMessage(ss.str(), IMGTYPE_PROFILE, EVENT_NOWORK);
+        return eCheckWorkResult::REFUSES;
+    }
+    return eCheckWorkResult::ACCEPTS;
+}
+
+PracticeJob::PracticeJob() : cBasicJob(JOB_TRAINING, "Training.xml") {
+
+}
+
+namespace {
+    Image_Types skill_to_image(SKILLS skill) {
+        switch (skill) {
+            case SKILL_STRIP:
+                return IMGTYPE_STRIP;
+            case SKILL_GROUP:
+                return IMGTYPE_GROUP;
+            case SKILL_BDSM:
+                return IMGTYPE_BDSM;
+            case SKILL_PERFORMANCE:
+                return IMGTYPE_MAST;
+            case SKILL_ANAL:
+                return IMGTYPE_ANAL;
+            case SKILL_NORMALSEX:
+                return IMGTYPE_SEX;
+            case SKILL_ORALSEX:
+                return IMGTYPE_ORAL;
+            case SKILL_TITTYSEX:
+                return IMGTYPE_TITTY;
+            case SKILL_LESBIAN:
+                return IMGTYPE_LESBIAN;
+            case SKILL_HANDJOB:
+                return IMGTYPE_HAND;
+            case SKILL_FOOTJOB:
+                return IMGTYPE_FOOT;
+            case SKILL_BEASTIALITY:
+                return IMGTYPE_BEAST;
+            default:
+                return IMGTYPE_PROFILE;
+        }
+    }
+}
+
+
+sWorkJobResult PracticeJob::DoWork(sGirl& girl, bool is_night) {
+    auto building = girl.m_Building;
+    sGirl* mistress = building->RequestInteraction(TrainingInteractionId);
+    SetSubstitution("mistress", mistress ? mistress->FullName() : "The Mistress");
+    girl.tiredness(2);
+
+    if(!mistress) {
+        add_text("no-mistress");
+        RandomSelector<SKILLS> selector;
+        SKILLS skills[] = {SKILL_NORMALSEX, SKILL_ANAL, SKILL_FOOTJOB, SKILL_HANDJOB, SKILL_LESBIAN,
+                           SKILL_TITTYSEX, SKILL_ORALSEX, SKILL_PERFORMANCE, SKILL_STRIP};
+        for(auto& skill : skills) {
+            if(girl.get_skill(skill) < 50 && girl.m_Building->is_sex_type_allowed(skill)) {
+                selector.process(&skill, 50 - girl.get_skill(skill));
+            }
+        }
+
+        girl.exp(5);
+        Image_Types image = IMGTYPE_MAST;
+        if(selector.selection()) {
+            auto target = *selector.selection();
+            image = skill_to_image(target);
+            if(girl.get_skill(target) < 50) {
+                girl.upd_skill(target, 1);
+                ss << "\n" << get_skill_name(target) << " + 1";
+            }
+        }
+
+        girl.AddMessage(ss.str(), image, is_night ? EVENT_NIGHTSHIFT : EVENT_DAYSHIFT);
+        return sWorkJobResult{false, 0, 0, 20};
+    } else {
+        mistress->tiredness(2);
+
+        // The mistress trains her. First, settle on a skill to train
+        RandomSelector<SKILLS> selector;
+        SKILLS skills[] = {SKILL_NORMALSEX, SKILL_ANAL, SKILL_BDSM, SKILL_BEASTIALITY, SKILL_FOOTJOB,
+                           SKILL_HANDJOB, SKILL_LESBIAN, SKILL_TITTYSEX, SKILL_ORALSEX, SKILL_GROUP,
+                           SKILL_PERFORMANCE, SKILL_STRIP};
+
+        for(auto& skill : skills) {
+            if(!girl.m_Building->is_sex_type_allowed(skill))
+                continue;
+
+            int my_value = girl.get_skill(skill);
+            int other_value = mistress->get_skill(skill);
+            float weight = 5;       // base chance
+            // is it important to train that value
+            if(my_value < 20)
+                weight += 5;
+            if(my_value < 50)
+                weight += 5;
+
+            // can I train that value well
+            if(my_value < other_value) {
+                weight += float(other_value - my_value) / 2;
+            }
+            selector.process(&skill, weight);
+        }
+
+        // this will always be non NULL, because SKILL_PERFORMANCE cannot be forbidden
+        SKILLS target = *selector.selection();
+        Image_Types image = skill_to_image(target);
+
+        SetSubstitution("skill", get_skill_name(target));
+        int my_value = girl.get_skill(target);
+        int other_value = mistress->get_skill(target);
+        if(my_value >= other_value) {
+            add_text("lack-of-skill");
+            girl.exp(5);
+            // if total skill < 50, we still get some update
+            if(girl.get_skill(target) < 50) {
+                girl.upd_skill(target, 1);
+                ss << "\n" << get_skill_name(target) << " + 1";
+            }
+            if(chance(5)) {
+                mistress->upd_skill(target, 1);
+            }
+        } else {
+            add_text(std::string("train.") + get_skill_name(target));
+            girl.exp(5);
+            int perf = mistress->job_performance(JOB_MISTRESS, false);
+            int min_gain = 1;
+            int max_gain = 3;
+            if(my_value < 20 && perf > 75) {
+                min_gain += 1;
+            }
+            if(other_value > 66 && my_value < 40 && perf > 100) {
+                min_gain += 1;
+            }
+
+            if(my_value < 40 && perf > 140) {
+                max_gain += 1;
+            }
+
+            if (girl.has_active_trait("Quick Learner"))     max_gain += 1;
+            else if (girl.has_active_trait("Slow Learner")) max_gain -= 1;
+
+            max_gain += std::min((other_value - my_value) / 10, 3);
+            int amount = uniform(min_gain, std::max(max_gain, min_gain));
+            girl.upd_skill(target, amount);
+            ss << "\n${name}'s " << get_skill_name(target) << " skill increased by " << amount << " points.";
+        }
+
+        girl.AddMessage(ss.str(), image, is_night ? EVENT_NIGHTSHIFT : EVENT_DAYSHIFT);
+        return sWorkJobResult{false, 0, 0, 20};
+    }
+}
+
+IGenericJob::eCheckWorkResult PracticeJob::CheckWork(sGirl& girl, bool is_night) {
+    if (girl.disobey_check(ACTION_WORKTRAINING, JOB_TRAINING))            // they refuse to work
+    {
+
+        sGirl* mistress = nullptr;
+        if(girl.is_slave() && (mistress = girl.m_Building->RequestInteraction(TrainingInteractionId))) {
+            SetSubstitution("mistress", mistress->FullName() );
+            add_text("refuse.forced");
+            // smaller changes than for dungeon torture, but still we should combine the code at some point
+            girl.obedience(2);
+            girl.upd_temp_stat(STAT_OBEDIENCE, 5);
+            girl.pcfear(1);
+            girl.pchate(1);
+            girl.spirit(-1);
+            girl.happiness(-2);
+            girl.tiredness(2);
+            girl.health(-2);
+
+            girl.AddMessage(ss.str(), IMGTYPE_TORTURE, EVENT_NOWORK);
+        } else {
+            add_text("refuse");
+            girl.AddMessage(ss.str(), IMGTYPE_PROFILE, EVENT_NOWORK);
+        }
+
+        return eCheckWorkResult::REFUSES;
+    }
+    return eCheckWorkResult::ACCEPTS;
+}
 
 class TrainingJob : public IGenericJob {
 public:
@@ -60,6 +297,8 @@ protected:
 
 private:
     int m_StartProgress = 0;
+
+    sGirl* m_Mistress = nullptr;
 };
 
 void TrainingJob::update_progress(sGirl& girl) {
@@ -114,6 +353,13 @@ IGenericJob::eCheckWorkResult TrainingJob::CheckWork(sGirl& girl, bool is_night)
         girl.FullJobReset(JOB_RESTING);
         girl.m_PrevWorkingDay = girl.m_WorkingDay = 0;
         return eCheckWorkResult::IMPOSSIBLE;    // not refusing
+    }
+
+    m_Mistress = girl.m_Building->RequestInteraction(TrainingInteractionId);
+    if(!m_Mistress) {
+        ss << "There is no Mistress available to train ${name}\n";
+        girl.AddMessage(ss.str(), IMGTYPE_PROFILE, EVENT_WARNING);
+        return eCheckWorkResult::IMPOSSIBLE;
     }
 
     return eCheckWorkResult::ACCEPTS;
@@ -500,4 +746,6 @@ void RegisterTrainingJobs(cJobManager& mgr) {
     mgr.register_job(std::make_unique<SoLesbian>());
     mgr.register_job(std::make_unique<SoBi>());
     mgr.register_job(std::make_unique<FakeOrg>());
+    mgr.register_job(std::make_unique<PracticeJob>());
+    mgr.register_job(std::make_unique<MistressJob>());
 }